feat: Refactor Etrian Calendar components and add registry features
- Remove unused EtrianItemSkeleton component.
- Update TodaysEtrianDateDisplay to use etrianNewYearsEve constant.
- Introduce color constants for birthday variants and month colors.
- Add date constants for Etrian months and New Year's Eve.
- Implement DateOfBirthBadge and AffiliationBadge components.
- Create BirthdayMessage component to display birthday notifications.
- Add BackupDialog and ConfirmDialog for managing Etrian data.
- Implement EditDialog for editing Etrian profiles.
- Create RegistryForm for adding new Etrians.
- Develop RegistryList to display Etrian entries with actions.
- Add hooks for managing Etrian registry in local storage.
- Define Zod schema for Etrian registration form validation.
- Update utility functions to reflect new constants structure.
