+++ src/app/(toys)/etrian-calendar/_constants/color.ts
@@ -1,0 +1,34 @@
+import {
+  EtrianMonthName,
+  EtrianNewYearsEveName,
+} from "@/app/(toys)/etrian-calendar/types/etrian";
+
+export type MonthColorKey = EtrianMonthName | EtrianNewYearsEveName | "default";
+
+export const BIRTHDAY_COLOR_VARIANTS = {
+  default: "bg-zinc-100 text-zinc-500 hover:bg-zinc-100",
+  kotei: "bg-sky-100 text-sky-700 hover:bg-sky-100",
+  sakura: "bg-pink-50 text-pink-500 hover:bg-pink-50",
+  koseki: "bg-green-100 text-lime-700 hover:bg-green-100",
+  johi: "bg-orange-100 text-red-700 hover:bg-orange-100",
+  rikka: "bg-slate-100 text-sky-600 hover:bg-slate-100",
+  monoka: "bg-rose-200 text-rose-700 hover:bg-rose-200",
+} as const;
+
+export const BIRTHDAY_CLASS_BY_MONTH: Record<MonthColorKey, string> = {
+  default: BIRTHDAY_COLOR_VARIANTS.default,
+  皇帝ノ月: BIRTHDAY_COLOR_VARIANTS.kotei,
+  笛鼠ノ月: BIRTHDAY_COLOR_VARIANTS.sakura,
+  天牛ノ月: BIRTHDAY_COLOR_VARIANTS.sakura,
+  王虎ノ月: BIRTHDAY_COLOR_VARIANTS.sakura,
+  素兎ノ月: BIRTHDAY_COLOR_VARIANTS.koseki,
+  虹竜ノ月: BIRTHDAY_COLOR_VARIANTS.koseki,
+  白蛇ノ月: BIRTHDAY_COLOR_VARIANTS.koseki,
+  風馬ノ月: BIRTHDAY_COLOR_VARIANTS.johi,
+  金羊ノ月: BIRTHDAY_COLOR_VARIANTS.johi,
+  飛猴ノ月: BIRTHDAY_COLOR_VARIANTS.johi,
+  火鳥ノ月: BIRTHDAY_COLOR_VARIANTS.rikka,
+  戌神ノ月: BIRTHDAY_COLOR_VARIANTS.rikka,
+  怒猪ノ月: BIRTHDAY_COLOR_VARIANTS.rikka,
+  鬼乎ノ日: BIRTHDAY_COLOR_VARIANTS.monoka,
+} as const;
