+++ src/app/(toys)/etrian-calendar/_features/registry/dialog/index.ts
@@ -1,0 +1,3 @@
+export { BackupDialog } from "./backup-dialog";
+export { ConfirmDialog } from "./confirm-dialog";
+export { EditDialog } from "./edit-dialog";
