+++ src/app/(toys)/etrian-calendar/_hooks/index.ts
@@ -1,0 +1,4 @@
+export {
+  ETRIAN_REGISTRY_STORAGE_KEY,
+  useEtrianRegistry,
+} from "./use-etrian-registry";
