+++ src/app/(toys)/etrian-calendar/_hooks/use-etrian-registry.ts
@@ -1,0 +1,77 @@
+import { useCallback, useEffect, useState } from "react";
+
+import { Etrian } from "@/app/(toys)/etrian-calendar/types/etrian";
+
+export const ETRIAN_REGISTRY_STORAGE_KEY = "etrianRegistry";
+
+type UseEtrianRegistryOptions = {
+  storageKey?: string;
+};
+
+type UseEtrianRegistryReturn = {
+  storedEtrians: Etrian[];
+  isLoaded: boolean;
+  addEtrian: (etrian: Etrian) => void;
+  updateEtrian: (etrian: Etrian) => void;
+  deleteEtrianById: (id: string) => void;
+  resetEtrians: () => void;
+};
+
+export function useEtrianRegistry(
+  options: UseEtrianRegistryOptions = {},
+): UseEtrianRegistryReturn {
+  const { storageKey = ETRIAN_REGISTRY_STORAGE_KEY } = options;
+
+  const [storedEtrians, setStoredEtrians] = useState<Etrian[]>([]);
+  const [isLoaded, setIsLoaded] = useState(false);
+
+  useEffect(() => {
+    if (typeof window === "undefined") {
+      return;
+    }
+
+    const data = window.localStorage.getItem(storageKey);
+    try {
+      setStoredEtrians(data ? (JSON.parse(data) as Etrian[]) : []);
+    } catch {
+      setStoredEtrians([]);
+    } finally {
+      setIsLoaded(true);
+    }
+  }, [storageKey]);
+
+  useEffect(() => {
+    if (!isLoaded || typeof window === "undefined") {
+      return;
+    }
+
+    window.localStorage.setItem(storageKey, JSON.stringify(storedEtrians));
+  }, [storedEtrians, isLoaded, storageKey]);
+
+  const addEtrian = useCallback((etrian: Etrian) => {
+    setStoredEtrians((prev) => [etrian, ...prev]);
+  }, []);
+
+  const updateEtrian = useCallback((updated: Etrian) => {
+    setStoredEtrians((prev) =>
+      prev.map((current) => (current.id === updated.id ? updated : current)),
+    );
+  }, []);
+
+  const deleteEtrianById = useCallback((id: string) => {
+    setStoredEtrians((prev) => prev.filter((etrian) => etrian.id !== id));
+  }, []);
+
+  const resetEtrians = useCallback(() => {
+    setStoredEtrians([]);
+  }, []);
+
+  return {
+    storedEtrians,
+    isLoaded,
+    addEtrian,
+    updateEtrian,
+    deleteEtrianById,
+    resetEtrians,
+  };
+}
